Better turn anticipation
The previous version would intercept the FDM course nicely, but
sometimes it ended up oscillating around the track rather than
permanently capturing it. Whether this is due to floating-point
inaccuracies or some deeper flaw I don't know; the braindead simple fix
is to just detect when we're close enough (< 0.1nm) to the target track,
and from then on just home in on the next waypoint.