Overhaul pitch trim mechanics
The CRJ uses both hstab incidence and elevator for trimming; within the
range of hstab incidence, pitch trim input maps to hstab incidence, but
once its limit is reached, elevator trim is applie on top.

Previously, this was achieved through some hacky Nasal override of the
elevatorTrim() function; however, this is problematic for two reasons:
the autopilot uses the elevator-trim input directly, which means it will
not go through the mapping logic, and turning off the autopilot in a
situation where there is strong elevator trim will leave the pilot with
insufficient elevator authority - e.g., during approach, the autopilot
will trim the aircraft up quite a bit, but the hstab incidence might not
be set accordingly, so when the pilot turns off the autopilot, the
extreme nose-up pitch means we cannot pull up any further, and the
Nasal logic makes it such that there is no way to turn the elevator trim
back down.

The second reason is that many joystick configurations will pop up a
message when the user changes pitch trim, and that message reports the
value of the elevator-pitch property, which is only changed after
exceeding the hstab incidence range. This is confusing, because now
inputting pitch trim will in fact change the pitch trim, but the popup
keeps reporting whatever elevator pitch is set.

So instead, we don't manipulate the elevator-trim input at all.
Instead, we add a property rule that maps it to two new properties:
fcs/elevator-trim and fcs/hstab-trim. These, then, are fed to the FDM.
However, because yasim uses the elevator-trim input for its solver, we
still keep that as an input for the elevator trim, and provide a
correction factor that cancels out the difference between elevator-trim
input and fcs/elevator-trim as calculated in the property rule.